// src/app/types/character.ts
export class Character {
  name: string;
  alliance?: string;
  thumbnail?: string;
  firstName?: string;
  skill?: string;
  stat?: string;
  flower?: string;
  gift?: string;
  lostItems?: string;
  trainings?: string;
  lectureQuestion1?: string;
  lectureAnswer1?: string;
  lectureQuestion2?: string;
  lectureAnswer2?: string;
  image?: string;
}